+++ src/db/models/new/entry/index.ts
@@ -2,23 +2,7 @@ import {Model} from '../../..';
 import {RevisionModel} from '../revision';
 import {WorkbookModel} from '../workbook';
 import {Favorite} from '../favorite';
-
-export enum EntryScope {
-    Connection = 'connection',
-    Dataset = 'dataset',
-    Widget = 'widget',
-    Dash = 'dash',
-    Folder = 'folder',
-    Config = 'config',
-    Pdf = 'pdf',
-}
-
-// "type" is a string field and value can be any string, this enum is only used for checks in code
-export enum EntryType {
-    File = 'file',
-    GsheetsV2 = 'gsheets_v2',
-    // types can be added as needed
-}
+import {EntryScope} from './types';
 
 export const EntryColumn = {
     Scope: 'scope',
