+++ src/services/new/entry/copy-entry-to-workbook.ts
@@ -2,7 +2,8 @@ import {AppError} from '@gravity-ui/nodekit';
 import {transaction} from 'objection';
 import {getId} from '../../../db';
 import OldEntry from '../../../db/models/entry';
-import {Entry, EntryScope} from '../../../db/models/new/entry';
+import {EntryScope} from '../../../db/models/new/entry/types';
+import {Entry} from '../../../db/models/new/entry';
 import {RevisionModel, RevisionModelColumn} from '../../../db/models/new/revision';
 import {SyncLinks} from '../../../types/models/link';
 import {US_ERRORS, BiTrackingLogs} from '../../../const';
