+++ src/services/new/workbook/get-workbook-content.ts
@@ -1,7 +1,7 @@
 import {makeSchemaValidator} from '../../../components/validation-schema-compiler';
 import {RETURN_NAVIGATION_COLUMNS, DEFAULT_PAGE, DEFAULT_PAGE_SIZE} from '../../../const';
 import Navigation from '../../../db/models/navigation';
-import {EntryScope} from '../../../db/models/new/entry';
+import {EntryScope} from '../../../db/models/new/entry/types';
 import Utils, {logInfo} from '../../../utils';
 import {UsPermission} from '../../../types/models';
 import {ServiceArgs} from '../types';
