+++ src/services/new/workbook/utils/get-entry-permissions-by-workbook.ts
@@ -3,7 +3,7 @@ import {Permissions} from '../../../../entities/workbook';
 import type {WorkbookInstance} from '../../../../registry/common/entities/workbook/types';
 import {Feature, isEnabledFeature} from '../../../../components/features';
 import type {EntryScope as EntryScopeType} from '../../../../types/models';
-import {EntryScope} from '../../../../db/models/new/entry';
+import {EntryScope} from '../../../../db/models/new/entry/types';
 
 export const getEntryPermissionsByWorkbook = ({
     ctx,
