+++ src/utils/utils.ts
@@ -3,17 +3,10 @@ const PowerRadix = require('power-radix');
 
 import {EntryScope, USAPIResponse} from '../types/models';
 
-import {
-    ID_VARIABLES,
-    CODING_BASE,
-    TRUE_FLAGS,
-    COPY_START,
-    COPY_END,
-    ENTRY_SCOPE,
-    ENTRY_TYPE,
-} from '../const';
+import {ID_VARIABLES, CODING_BASE, TRUE_FLAGS, COPY_START, COPY_END} from '../const';
 
-import {Entry} from '../db/models/new/entry';
+import type {Entry} from '../db/models/new/entry';
+import {EntryScope as EntryScopeEnum, EntryType} from '../db/models/new/entry/types';
 
 const PROFILES: {
     [key: string]: any;
@@ -347,11 +340,12 @@ export class Utils {
     }
 
     static checkFileConnectionsExistence(entries: Entry[]) {
-        const fileConnectionTypes: string[] = [ENTRY_TYPE.File, ENTRY_TYPE.GsheetsV2];
+        const fileConnectionTypes: string[] = [EntryType.File, EntryType.GsheetsV2];
 
         return entries.some((entry) => {
             return (
-                entry.scope === ENTRY_SCOPE.Connection && fileConnectionTypes.includes(entry.type)
+                entry.scope === EntryScopeEnum.Connection &&
+                fileConnectionTypes.includes(entry.type)
             );
         });
     }
